Assignment_2_1.00
All methods and operations have been implemented, and the game runs fine. It's important to note, though, that its a bit difficult to compile the program using a terminal, so String[] args2 must be modified with a user's desired input before running the game. the game works as it should, but the computer AI seems to struggle on larger boards, or ones with a winning sequence that is shorter than the board size.