Assignment_4_0.40
The BinarySearchTree class is now complete, and should operate properly, assuming I've inserted/removed things based on the correct values, since the keys are objects, I've resorted to simply parseInt()'ing the key's label, and comparing the integers of what is being inserted/removed to the current node we're looking at (a.k.a. p, for some reason).  BSTDictionary and Interface still need to be coded, the former should be completed by 0.50.