Assignment_4_0.65
Interface groundwork has been placed, lots of changes need to be made to the BST class, since ordering things is incorrect and get()/insert()/remove() all traverse incorrectly.  Additionally, when reorganizing a tree after removal of an internal node, a successor can be used, otherwise a predecessor must be used.