Assignment_4_0.910
Further debugging has shown that Key.compareTo() was not functioning properly because the String.compareTo() method was expected to return -1 when it simply returns any negative number. this issue has been resolved and the BST is no longer operating like a LL, successor and predecessor seem to be semi-functional but lack the ability to find the correct node when travelling upwards towards the root, or at least that is what I suppose is occuring.