Assignment_4_0.950
All BST functions are operating successfully now. The main problems experienced was a reversed integer value in both successor/predecessor methods, and an issue where the wrapper predecessor/successor methods were both calling successor only, which is a frustrating oversight to say the least. compareTo() needed to be altered again since "label == k.getLabel()" was never true because it was most likely comparing the two objects to see if the match exactly, and not whether the strings themselves match.